// src/Dogger/wwwroot/src/api/openapi/models/ExposedPort.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    SocketProtocol,
    SocketProtocolFromJSON,
    SocketProtocolFromJSONTyped,
    SocketProtocolToJSON,
} from './';

/**
 * 
 * @export
 * @interface ExposedPort
 */
export interface ExposedPort {
    /**
     * 
     * @type {number}
     * @memberof ExposedPort
     */
    port?: number;
    /**
     * 
     * @type {SocketProtocol}
     * @memberof ExposedPort
     */
    protocol?: SocketProtocol;
}

export function ExposedPortFromJSON(json: any): ExposedPort {
    return ExposedPortFromJSONTyped(json, false);
}

export function ExposedPortFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExposedPort {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'port': !exists(json, 'port') ? undefined : json['port'],
        'protocol': !exists(json, 'protocol') ? undefined : SocketProtocolFromJSON(json['protocol']),
    };
}

export function ExposedPortToJSON(value?: ExposedPort | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'port': value.port,
        'protocol': SocketProtocolToJSON(value.protocol),
    };
}