// src/Dogger/wwwroot/src/api/openapi/models/LogsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LogsResponse
 */
export interface LogsResponse {
    /**
     * 
     * @type {string}
     * @memberof LogsResponse
     */
    containerImage?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LogsResponse
     */
    containerId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LogsResponse
     */
    logs?: string | null;
}

export function LogsResponseFromJSON(json: any): LogsResponse {
    return LogsResponseFromJSONTyped(json, false);
}

export function LogsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LogsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'containerImage': !exists(json, 'containerImage') ? undefined : json['containerImage'],
        'containerId': !exists(json, 'containerId') ? undefined : json['containerId'],
        'logs': !exists(json, 'logs') ? undefined : json['logs'],
    };
}

export function LogsResponseToJSON(value?: LogsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'containerImage': value.containerImage,
        'containerId': value.containerId,
        'logs': value.logs,
    };
}